--- planet.rb
@@ -1,17 +1,0 @@
-class Planet
-
-  attr_reader :name, :color, :mass_kg, :distance_from_sun_km, :fun_fact
-
-  def initialize(name, color, mass_kg, distance_from_sun_km, fun_fact)
-    @name = name
-    @color = color
-    @mass_kg = mass_kg
-    @distance_from_sun_km = distance_from_sun_km
-    @fun_fact = fun_fact
-  end
-
-  def summary
-    return "Planet: #{name} \n1. Color: #{color} \n2. Mass (kg): #{mass_kg} \n3. Distance from sun (km): #{distance_from_sun_km} \n4. Fun fact: #{fun_fact}"
-  end
-
-end
